+++ frontend/src/views/system/action-permission/modules/action-permission-search.vue
@@ -1,18 +1,16 @@
 <template>
   <ArtSearchBar
-    ref="searchBarRef"
     v-model="formData"
     :items="formItems"
-    :rules="rules"
     :showExpand="true"
-    :defaultExpanded="false"
+    :defaultExpanded="true"
     @search="handleSearch"
     @reset="handleReset"
   />
 </template>
 
 <script setup lang="ts">
-  import { computed, ref } from 'vue'
+  import { computed } from 'vue'
   import type { FormItem } from '@/components/core/forms/art-form/index.vue'
 
   interface SearchForm {
@@ -43,14 +41,11 @@
 
   const props = defineProps<Props>()
   const emit = defineEmits<Emits>()
-  const searchBarRef = ref()
 
   const formData = computed({
     get: () => props.modelValue,
     set: (val) => emit('update:modelValue', val)
   })
-
-  const rules = {}
 
   const formItems = computed<FormItem[]>(() => [
     {
@@ -126,10 +121,7 @@
     }
   ])
 
-  const handleSearch = async () => {
-    await searchBarRef.value?.validate()
-    emit('search')
-  }
+  const handleSearch = () => emit('search')
 
   const handleReset = () => emit('reset')
 </script>