+++ frontend/src/views/system/menu/modules/menu-search.vue
@@ -1,16 +1,16 @@
 <template>
   <ArtSearchBar
-    ref="searchBarRef"
     v-model="formData"
     :items="formItems"
-    :showExpand="false"
+    :showExpand="true"
+    :defaultExpanded="true"
     @reset="handleReset"
     @search="handleSearch"
   />
 </template>
 
 <script setup lang="ts">
-  import { computed, ref } from 'vue'
+  import { computed } from 'vue'
   import type { FormItem } from '@/components/core/forms/art-form/index.vue'
 
   interface SearchForm {
@@ -30,7 +30,6 @@
 
   const props = defineProps<Props>()
   const emit = defineEmits<Emits>()
-  const searchBarRef = ref()
 
   const formData = computed({
     get: () => props.modelValue,
@@ -58,10 +57,7 @@
     }
   ])
 
-  const handleSearch = async () => {
-    await searchBarRef.value?.validate()
-    emit('search')
-  }
+  const handleSearch = () => emit('search')
 
   const handleReset = () => emit('reset')
 </script>